refactor: 사용하지 않는 파일 삭제 및 파일 이동
