--- android/app/src/main/java/com/yagubogu/presentation/util/ScrollToTop.kt
@@ -1,5 +1,0 @@
-package com.yagubogu.presentation.util
-
-fun interface ScrollToTop {
-    fun scrollToTop()
-}
