--- android/app/src/main/java/com/yagubogu/ui/common/component/shape/ShapeUtil.kt
@@ -1,12 +1,0 @@
-package com.yagubogu.ui.common.component.shape
-
-data class Point(
-    val x: Float,
-    val y: Float,
-)
-
-enum class CornerStyle {
-    ROUNDED,
-    INNER_ROUNDED,
-    CUT,
-}
